requirements/requirements.txt
alembic==1.10.3
anyio==3.7.1
asn1crypto==1.5.1
astroid==2.15.6
cbor2==5.4.6
cffi==1.15.1
click==8.1.4
cryptography==41.0.2
dill==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
passlib==1.7.4
platformdirs==3.8.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
pylint==2.17.2
pyOpenSSL==23.2.0
pytest==7.3.1
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomlkit==0.11.8
typing_extensions==4.7.1
uvicorn==0.21.1
webauthn==1.8.1
wrapt==1.15.0
